cv_tailor_project/requirements.txt
python-dotenv
PyPDF2
python-docx
google-genai
Flask
jobspy
reportlab
